requirements.txt
fastapi
uvicorn[standard]
pyjwt
mongoengine
python-multipart
ujson